requirements.txt
panel
tornado
fastapi
authlib
python-dotenv
uvicorn
httpx
itsdangerous